move FieldSet out of subdirectory
