+++ frontend/src/pages/projects/[project]/settings.tsx
@@ -7,8 +7,8 @@ import { useRouter } from "next/router"
 import { SWRConfig } from "swr"
 
 import AsyncButton from "../../../components/AsyncButton"
+import FieldSet from "../../../components/FieldSet"
 import Footer from "../../../components/Footer"
-import { FieldSet } from "../../../components/form/FieldSet"
 import ImageInput from "../../../components/ImageInput"
 import Nav from "../../../components/Nav"
 import PageTitle from "../../../components/PageTitle"
